+++ src/noahtheduke/splint/rules/style/prefer_clj_string.clj
@@ -1,0 +1,71 @@
+; This Source Code Form is subject to the terms of the Mozilla Public
+; License, v. 2.0. If a copy of the MPL was not distributed with this
+; file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+(ns ^:no-doc noahtheduke.splint.rules.style.prefer-clj-string
+  (:require
+    [noahtheduke.spat.pattern :as p]
+    [noahtheduke.splint.diagnostic :refer [->diagnostic]]
+    [noahtheduke.splint.rules :refer [defrule postwalk-splicing-replace]]))
+
+(set! *warn-on-reflection* true)
+
+(defn upper-case?? [sexp]
+  (case (name sexp)
+    (".toUpperCase" "upper-case") true
+    false))
+
+(defn lower-case?? [sexp]
+  (case (name sexp)
+    (".toLowerCase" "lower-case") true
+    false))
+
+(defn to-str?? [sexp]
+  (case (name sexp)
+    (".toString" "str") true
+    false))
+
+(def capitalize??
+  {:pattern (p/pattern '(%to-str?? (%upper-case?? (subs ?s 0 1))
+                                   (%lower-case?? (subs ?s 1))))
+   :replace '(clojure.string/capitalize ?s)})
+
+(def reverse??
+  {:pattern (p/pattern '(%to-str?? (.reverse (StringBuilder. ?s))))
+   :replace '(clojure.string/reverse ?s)})
+
+(def interop->clj-string
+  '{.contains clojure.string/includes?
+    .endsWith clojure.string/ends-with?
+    .replace clojure.string/replace
+    .split clojure.string/split
+    .startsWith clojure.string/starts-with?
+    .toLowerCase clojure.string/lower-case
+    ; TODO: Need a way to flag overlapping rules.
+    #_#_.toString clojure.core/str
+    .toUpperCase clojure.string/upper-case
+    .trim clojure.string/trim})
+
+(defrule style/prefer-clj-string
+  "Prefer clojure.math to interop.
+
+  Examples:
+
+  # bad
+  (.toUpperCase \"hello world\")
+
+  # good
+  (clojure.string/upper-case \"hello world\")
+  "
+  {:pattern '(%symbol?%-?sym &&. ?args)
+   :message "Use the `clojure.string` function instead of interop."
+   :on-match (fn [ctx rule form {:syms [?sym ?args]}]
+               (if-let [binds ((:pattern capitalize??) form)]
+                 (let [new-form (postwalk-splicing-replace binds (:replace capitalize??))]
+                   (->diagnostic rule form {:replace-form new-form}))
+                 (if-let [binds ((:pattern reverse??) form)]
+                   (let [new-form (postwalk-splicing-replace binds (:replace reverse??))]
+                     (->diagnostic rule form {:replace-form new-form}))
+                   (when-let [replacement (interop->clj-string ?sym)]
+                     (let [new-form (apply list replacement ?args)]
+                       (->diagnostic rule form {:replace-form new-form}))))))})